moving lib into private folder
