+++ cortex/_lib/data/noise.py
