+++ setup.py
@@ -5,11 +5,11 @@ setup(name='cortex',
       description='A library for wrapping your pytorch code',
       author='R Devon Hjelm',
       author_email='erroneus@gmail.com',
-      packages=['lib'],
+      packages=['cortex'],
       install_requires=[
         'imageio', 'matplotlib', 'progressbar2', 'scipy', 'sklearn', 'torchvision', 'visdom', 'pyyaml'],
       entry_points={
         'console_scripts': [
-            'cortex=lib:main']
+            'cortex=cortex.main:main']
       },
       zip_safe=False)
